requirements.txt
streamlit
opencv-python
pytesseract
pandas
numpy
openai>=1.0.0
Pillow